Addresses #193. Added objective-c ACSE, DIMSE, stream reader/writer, dicomdir, dictionary

+++ library/include/imebra/objectivec/imebra_tagId.h
@@ -48,7 +48,7 @@ class TagId;
 
     @property (readonly) unsigned int groupOrder;
 
-    @property (readonly) unsigned int tagId;
+    @property (readonly) unsigned short tagId;
 
 @end
